refactor(ea): converge SysML hand-seed views onto one canonical applier (Parity Engine)
Five copies of the SysML create/update + relationship + view upsert logic collapse to one. The shared applySysmlModel now lives in packages/db (seed-time seeders cannot import apps/web — the workspace dependency only runs one way), exported via the @dpf/db/sysml-model-seed subpath; apps/web re-exports it under the stable ./sysml-model-seed path the extractors already use. The runtime auto-extractors (MCP authority, coworker workforce, value streams) already used the apps/web copy; the three hand-authored current-state views (AI cockpit, agent authority, data authority) now build a desired model and hand it to the same applier instead of carrying their own upsert/rel/view boilerplate.

Extended the applier so it can absorb the hand-seeds without losing fidelity: per-element provenance (architect-authored elements stay 'architect'; otherwise 'deterministic' — the parity gate depends on this distinction), an optional conformance-issue pass (guarded, so views that file none never touch the table), and a configurable lifecycleStage (the cockpit is design-stage, the others production).

Source-local: 17 packages/db tests (canonical applier incl. new provenance/conformance/lifecycle cases + all three converged views, every original assertion preserved) and 74 apps/web lib/ea tests pass; both packages typecheck clean (modulo the worktree-only @dpf/db module-resolution artifact, which resolves in CI). Branched fresh off origin/main.
