+++ packages/db/src/index.ts
@@ -134,6 +134,16 @@ export {
 } from "./schema-source";
 export { DATA_MODEL_MIRROR_TASK_ID } from "./data-model-mirror-config";
 export { SYSML_PROJECTION_TASK_ID } from "./sysml-projection-config";
+// Canonical SysML projection applier — shared by seed-time views (packages/db
+// seed-ea-sysml-*.ts) and runtime extractors (apps/web/lib/ea, via this barrel).
+export {
+  applySysmlModel,
+  type SysmlDesiredElement,
+  type SysmlDesiredRel,
+  type SysmlDesiredConformanceIssue,
+  type SysmlDesiredModel,
+  type SysmlSeedResult,
+} from "./sysml-model-seed";
 export {
   buildDiscoveredKey,
   buildInventoryEntityKey,
