+++ packages/db/src/seed-ea-sysml-ai-cockpit.ts
@@ -20,10 +20,9 @@
 // deferred follow-up thread. This module is unit-tested for idempotency with a
 // mocked client; it is not asserted to have run against a live database here.
 
-import { prisma } from "./client.js";
 import type { Prisma } from "../generated/client/client";
+import { applySysmlModel, type SysmlDesiredElement, type SysmlDesiredModel, type SysmlDesiredRel } from "./sysml-model-seed.js";
 
-const NOTATION_SLUG = "sysml2";
 const KEY = (suffix: string) => `sysml:aic:${suffix}`;
 
 type ElementSeed = {
@@ -100,129 +99,43 @@ const PACKAGE: ElementSeed = {
 
 const VIEW_NAME = "AI Cockpit — Requirements & Verification";
 
+function buildAiCockpitModel(): SysmlDesiredModel {
+  const all: ElementSeed[] = [PACKAGE, ...REQUIREMENTS, ...CONSTRAINTS, ...PARTS, ...INTERFACES, ...VERIFICATIONS];
+  const elements: SysmlDesiredElement[] = all.map((seed) => ({
+    sysmlKey: KEY(seed.key),
+    typeSlug: seed.typeSlug,
+    name: seed.name,
+    description: seed.description,
+    properties: (seed.properties ?? {}) as Record<string, unknown>,
+  }));
+
+  const relationships: SysmlDesiredRel[] = [];
+  for (const seed of all) {
+    if (seed.key !== PACKAGE.key) relationships.push({ fromKey: KEY(PACKAGE.key), toKey: KEY(seed.key), relSlug: "contains" });
+  }
+  for (const p of PARTS) for (const reqKey of p.satisfies) relationships.push({ fromKey: KEY(p.key), toKey: KEY(reqKey), relSlug: "satisfies" });
+  for (const v of VERIFICATIONS) for (const reqKey of v.verifies) relationships.push({ fromKey: KEY(v.key), toKey: KEY(reqKey), relSlug: "verifies" });
+  for (const ref of CONSTRAINT_REFINES) relationships.push({ fromKey: KEY(ref.from), toKey: KEY(ref.to), relSlug: "refines" });
+
+  return {
+    elements,
+    relationships,
+    elementTypeSlugs: ["package", "requirement", "constraint", "part_definition", "interface_definition", "verification_case"],
+    relTypeSlugs: ["contains", "satisfies", "verifies", "refines"],
+    view: {
+      name: VIEW_NAME,
+      description: "Requirements, the parts that satisfy them, and the verification cases that prove them, for the AI Cockpit model.",
+      viewpointName: "Systems Requirements & Verification",
+      scopeRef: "ai-cockpit",
+    },
+    lifecycleStage: "design",
+  };
+}
+
 export async function seedEaSysmlAiCockpit(): Promise<void> {
-  const notation = await prisma.eaNotation.findUnique({ where: { slug: NOTATION_SLUG }, select: { id: true } });
-  if (!notation) {
-    console.warn(`Skipping AI-cockpit SysML model: notation "${NOTATION_SLUG}" not seeded`);
-    return;
+  const model = buildAiCockpitModel();
+  const r = await applySysmlModel(model);
+  if (r.status !== "skipped") {
+    console.log(`Seeded AI Cockpit SysML model: ${model.elements.length} elements + view "${VIEW_NAME}"`);
   }
-  const notationId = notation.id;
-
-  // Resolve element + relationship type ids for the slugs this model uses.
-  const typeSlugs = ["package", "requirement", "constraint", "part_definition", "interface_definition", "verification_case"];
-  const etId = new Map<string, string>();
-  for (const slug of typeSlugs) {
-    const et = await prisma.eaElementType.findUniqueOrThrow({
-      where: { notationId_slug: { notationId, slug } },
-      select: { id: true },
-    });
-    etId.set(slug, et.id);
-  }
-  const relSlugs = ["contains", "satisfies", "verifies", "refines"];
-  const rtId = new Map<string, string>();
-  for (const slug of relSlugs) {
-    const rt = await prisma.eaRelationshipType.findUniqueOrThrow({
-      where: { notationId_slug: { notationId, slug } },
-      select: { id: true },
-    });
-    rtId.set(slug, rt.id);
-  }
-
-  // Upsert an element by infraCiKey (no natural unique key on EaElement).
-  const elementId = new Map<string, string>(); // model key → EaElement.id
-  async function upsertElement(seed: ElementSeed): Promise<string> {
-    const infraCiKey = KEY(seed.key);
-    const elementTypeId = etId.get(seed.typeSlug);
-    if (!elementTypeId) throw new Error(`AI-cockpit seed: unknown sysml2 element type "${seed.typeSlug}"`);
-    const data = {
-      elementTypeId,
-      name: seed.name,
-      description: seed.description,
-      properties: seed.properties ?? {},
-      lifecycleStage: "design",
-      lifecycleStatus: "active",
-      infraCiKey,
-    };
-    const existing = await prisma.eaElement.findFirst({ where: { infraCiKey }, select: { id: true } });
-    const row = existing
-      ? await prisma.eaElement.update({ where: { id: existing.id }, data, select: { id: true } })
-      : await prisma.eaElement.create({ data, select: { id: true } });
-    elementId.set(seed.key, row.id);
-    return row.id;
-  }
-
-  // Upsert a relationship (no natural unique key on EaRelationship).
-  async function upsertRel(fromKey: string, toKey: string, relSlug: string): Promise<void> {
-    const fromElementId = elementId.get(fromKey);
-    const toElementId = elementId.get(toKey);
-    const relationshipTypeId = rtId.get(relSlug);
-    if (!fromElementId || !toElementId || !relationshipTypeId) {
-      console.warn(`AI-cockpit seed: skipping rel ${fromKey} -[${relSlug}]-> ${toKey} (unresolved)`);
-      return;
-    }
-    const existing = await prisma.eaRelationship.findFirst({
-      where: { fromElementId, toElementId, relationshipTypeId },
-      select: { id: true },
-    });
-    if (!existing) {
-      await prisma.eaRelationship.create({
-        data: { fromElementId, toElementId, relationshipTypeId, notationSlug: NOTATION_SLUG, properties: {} },
-      });
-    }
-  }
-
-  // 1. Elements.
-  await upsertElement(PACKAGE);
-  for (const r of REQUIREMENTS) await upsertElement(r);
-  for (const c of CONSTRAINTS) await upsertElement(c);
-  for (const p of PARTS) await upsertElement(p);
-  for (const i of INTERFACES) await upsertElement(i);
-  for (const v of VERIFICATIONS) await upsertElement(v);
-  const elementCount = elementId.size;
-
-  // 2. Relationships.
-  // package contains every other element.
-  for (const key of elementId.keys()) {
-    if (key !== PACKAGE.key) await upsertRel(PACKAGE.key, key, "contains");
-  }
-  // parts satisfy requirements.
-  for (const p of PARTS) for (const reqKey of p.satisfies) await upsertRel(p.key, reqKey, "satisfies");
-  // verification cases verify requirements.
-  for (const v of VERIFICATIONS) for (const reqKey of v.verifies) await upsertRel(v.key, reqKey, "verifies");
-  // constraint refines requirements.
-  for (const ref of CONSTRAINT_REFINES) await upsertRel(ref.from, ref.to, "refines");
-
-  // 3. View: place the model under the "Systems Requirements & Verification" viewpoint.
-  const viewpoint = await prisma.viewpointDefinition.findUnique({
-    where: { name: "Systems Requirements & Verification" },
-    select: { id: true },
-  });
-  let view = await prisma.eaView.findFirst({
-    where: { notationId, name: VIEW_NAME },
-    select: { id: true },
-  });
-  if (!view) {
-    view = await prisma.eaView.create({
-      data: {
-        notationId,
-        name: VIEW_NAME,
-        description: "Requirements, the parts that satisfy them, and the verification cases that prove them, for the AI Cockpit model.",
-        layoutType: "graph",
-        scopeType: "custom",
-        scopeRef: "ai-cockpit",
-        viewpointId: viewpoint?.id ?? null,
-        status: "draft",
-      },
-      select: { id: true },
-    });
-  }
-  for (const id of elementId.values()) {
-    await prisma.eaViewElement.upsert({
-      where: { viewId_elementId: { viewId: view.id, elementId: id } },
-      update: {},
-      create: { viewId: view.id, elementId: id, mode: "existing" },
-    });
-  }
-
-  console.log(`Seeded AI Cockpit SysML model: ${elementCount} elements + view "${VIEW_NAME}"`);
 }
